+++ assets/init.js
@@ -1,0 +1,2 @@
+ace.config.set("basePath", "../assets/ace_editor");
+import {Editor} from editor.js
